modules/whisper-service/requirements.txt

# Async and utilities (proven working)
asyncio-mqtt>=0.16.2
audioread>=3.0.0
faster-whisper>=0.9.0  # Keep for compatibility
# Core Flask and web dependencies (proven working)
Flask>=2.0.0
Flask-CORS>=3.0.0
Flask-SocketIO>=5.0.0
hiredis>=2.0.0
librosa>=0.11.0

# Audio analysis and visualization (proven working)
matplotlib>=3.5.0

# Speech Enhancement
noisereduce>=2.0.0

# Audio processing (proven working)
numpy>=2.3.4
onnxruntime>=1.12.0  # Keep for compatibility

# Whisper and speech processing - Phase 2: PyTorch-based
openai-whisper>=20231117  # PyTorch Whisper implementation
pandas>=2.0.0

# Speaker Diarization and Audio Enhancement Dependencies
pyannote.audio>=3.1.0  # Enabled for Phase 2
pydub>=0.25.1
python-dotenv>=1.1.1
python-socketio>=5.14.2

# Audio feature extraction
python_speech_features>=0.6.0

# Redis for caching and session management
redis>=6.4.0
requests>=2.31.0
resemblyzer>=0.1.1
scikit-learn>=1.0.0
scipy>=1.16.2

# Real-time audio processing (may need system audio libraries)
# pyaudio>=0.2.11

# Additional dependencies for speaker diarization
seaborn>=0.11.0
silero-vad>=4.0.0
sounddevice>=0.4.0
soundfile>=0.13.1

# Advanced Audio Processing (optional for NPU)
speechbrain>=0.5.12
tiktoken>=0.5.0  # Token estimation for domain prompting
torch>=2.9.0  # PyTorch for GPU acceleration (latest)
torchaudio>=2.9.0  # Audio processing with PyTorch

# Enhanced audio processing
webrtcvad>=2.0.10

# WebSocket and networking (proven working)
websockets>=15.0.1
